Style improvements to make_release

+++ make_release.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-autoreconf
 make clean >/dev/null
 
 wget -q -O - http://downloads.sourceforge.net/project/samtools/samtools/1.1/samtools-1.1.tar.bz2 | tar -xj
@@ -17,7 +16,7 @@ echo -e "Bootstrap and configure"
 ./bootstrap && ./configure SAMTOOLS="${PWD}/samtools-1.1/" HTSLIB="${PWD}/samtools-1.1/htslib-1.1/" &> log_configure.log
 
 if [ "$(uname)" == 'Darwin' ]; then
-    echo -e "Making objects"
+    echo -e "Making objects for MacOsX"
     make -j 4 &> log_make_lofreq.log
     pushd src/lofreq
     rm lofreq
@@ -34,13 +33,14 @@ if [ "$(uname)" == 'Darwin' ]; then
     popd
 elif [ "$(uname)" == 'Linux' ]; then
     # assuming GIS setup with locally statically compiled libz
-    echo -e "Making objects"
+    echo -e "Making objects for linux"
     make -j 4 &> log_make_lofreq.log
 else
     echo "Unknown system: $(uname)" 1>&2
     exit 1
 fi
 
+echo -e "Objects done, now packaging"
 # all systems:
 version=$(grep '^AC_INIT' configure.ac | cut -f 2 -d , | tr -d '[\[\]] ')
 reldir=lofreq_star-$version
@@ -48,8 +48,7 @@ test -d "$reldir" && exit 1
 mkdir "$reldir"
 # use tar to preserve directory structure as if untouched src tree
 # but removing unwanted stuff
-tar c src/cdflib90.README src/lofreq/lofreq $(find src/scripts/ -name \*py -or -name \*py.README | grep -v '/build/' | grep -v setup) | tar x --strip-components 1 -C $reldir || exit 1
+tar c src/cdflib90.README src/lofreq/lofreq "$(find src/scripts/ -name \*py -or -name \*py.README | grep -v '/build/' | grep -v setup)" | tar x --strip-components 1 -C "$reldir" || exit 1
 cp README.md LICENSE "$reldir"
-tar cvzf $(basename $reldir).tgz $reldir && rm -rf $reldir/* && rmdir $reldir
-
+tar cvzf "$(basename $reldir)".tgz "$reldir" && rm -rf "${reldir:?}/"* && rmdir "$reldir"
 echo "release in $(basename $reldir).tgz. rename to architecture. unpack and test" 1>&2
